requirements.txt
customtkinter
matplotlib
pymongo
redis
neo4j
networkx
tksheet
cassandra-driver

